NF: Move Tasks in classes where they are used
