+++ AnkiDroid/src/main/java/com/ichi2/anki/Reviewer.java
@@ -53,14 +53,9 @@ import com.ichi2.anki.reviewer.ReviewerUi;
 import com.ichi2.anki.workarounds.FirefoxSnackbarWorkaround;
 import com.ichi2.async.CollectionTask;
 import com.ichi2.anki.reviewer.ActionButtons;
-import com.ichi2.async.task.AnswerCard;
-import com.ichi2.async.task.BuryCard;
-import com.ichi2.async.task.BuryNote;
 import com.ichi2.async.task.DismissMulti;
 import com.ichi2.async.task.RescheduleCards;
 import com.ichi2.async.task.ResetCards;
-import com.ichi2.async.task.SuspendCard;
-import com.ichi2.async.task.SuspendNote;
 import com.ichi2.compat.CompatHelper;
 import com.ichi2.libanki.Collection;
 import com.ichi2.libanki.Collection.DismissType;
